+++ src/pos/model/__init__.py
@@ -3,3 +3,4 @@
 from .abltagger import *
 from .decoders import *
 from .embeddings import *
+from .impl import *
